♻️ refactor(app): restructure component organization
- update component import paths to match new directory structure
- convert CVPage component from class to functional component

🔥 remove(cv-page): delete deprecated CVPage component

- remove outdated CVPage component logic and JSX structure

✨ feat(components): add hero, projects, skills, contact, and footer

- create new functional components for modular application structure
- enhance app maintainability and reusability with component-based design

💄 style(cv-page): apply container centering and padding in Tailwind config

- optimize layout by applying centered container with responsive padding
- improve visual consistency across different screen sizes
