+++ src/components/footer.tsx
@@ -1,0 +1,10 @@
+export default function Footer() {
+	return (
+		<footer className="text-center text-gray-400 mt-16">
+			<p>
+				Copyleft | William Ghazarian {new Date().getFullYear()} | No rights
+				reserved
+			</p>
+		</footer>
+	);
+}
